feat(apollo,look&feel): composant content item duo
