+++ client/apollo/css/src/List/ContentItemDuo/ContentItemDuoApollo.scss
@@ -1,0 +1,11 @@
+@use "./ContentItemDuoCommon";
+
+.af-content-item-duo {
+  &__label {
+    --content-duo-label-color: var(--neutral-80);
+  }
+
+  &__value {
+    --content-duo-label-color: var(--neutral-100);
+  }
+}
