+++ client/apollo/css/src/List/ContentItemDuo/ContentItemDuoLF.scss
@@ -1,0 +1,11 @@
+@use "./ContentItemDuoCommon";
+
+.af-content-item-duo {
+  &__label {
+    --content-duo-label-color: var(--color-gray-700);
+  }
+
+  &__value {
+    --content-duo-label-color: var(--color-gray-900);
+  }
+}
