+++ client/apollo/css/src/apollo.scss
@@ -22,6 +22,7 @@
 @use "./Form/DateInput/DateInputApollo";
 @use "./TimelineVertical/TimelineVerticalApollo";
 @use "./Heading/HeadingApollo";
+@use "./List/ContentItemDuo/ContentItemDuoApollo";
 @use "./Form/Checkbox/Checkbox/CheckboxApollo";
 @use "./Form/Checkbox/CheckboxText/CheckboxTextApollo";
 @use "./Form/Checkbox/CheckboxCard/CheckboxCardApollo";
