+++ client/apollo/css/src/look-and-feel.scss
@@ -16,6 +16,7 @@
 @use "./Message/MessageLF";
 @use "./Tag/TagLF";
 @use "./CardMessage/CardMessageLF";
+@use "./List/ContentItemDuo/ContentItemDuoLF";
 @use "./ProgressBar/ProgressBarLF";
 @use "./ProgressBarGroup/ProgressBarGroupLF";
 @use "./Stepper/StepperLF";
