+++ client/apollo/react/src/List/ContentItemDuo/ContentItemDuoApollo.tsx
@@ -1,0 +1,3 @@
+import "@axa-fr/design-system-apollo-css/dist/List/ContentItemDuo/ContentItemDuoApollo.scss";
+
+export { ContentItemDuo } from "./ContentItemDuoCommon";
