+++ client/apollo/react/src/List/ContentItemDuo/ContentItemDuoLF.tsx
@@ -1,0 +1,3 @@
+import "@axa-fr/design-system-apollo-css/dist/List/ContentItemDuo/ContentItemDuoLF.scss";
+
+export { ContentItemDuo } from "./ContentItemDuoCommon";
