--- client/look-and-feel/react/src/List/ContentItemDuo/index.ts
@@ -1,3 +1,0 @@
-import "@axa-fr/design-system-look-and-feel-css/dist/List/ContentItemDuo/ContentItemDuo.scss";
-
-export { ContentItemDuo } from "./ContentItemDuo";
